# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        example: doctor@clinica.com
        type: string
      first_name:
        example: Juan
        type: string
      last_name:
        example: Pérez
        type: string
      password:
        example: password123
        type: string
      phone:
        example: "+593999999999"
        type: string
      role:
        example: doctor
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: zensoftsas@gmail.com
    name: Zensoft Team
    url: https://github.com/zensoftsas/sistema-reservas
  description: |
    API completa para gestión de clínica médica con sistema de citas, servicios, horarios personalizados y analytics.

    Características principales:
    - Autenticación JWT con roles (admin, doctor, patient)
    - Sistema de citas médicas con validaciones
    - Servicios médicos con precios y duraciones
    - Horarios personalizados por doctor
    - Cálculo dinámico de slots disponibles
    - Analytics y reportes para administradores
    - Notificaciones por email (SendGrid)
    - Recordatorios automáticos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sistema de Reservas - Clínica Internacional API
  version: "1.0"
paths:
...
  /api/users:
    post:
      consumes:
      - application/json
      description: Crear un nuevo usuario (admin, doctor o patient)
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Crear usuario
      tags:
      - Users
  /api/users/me:
    get:
      consumes:
      - application/json
      description: Retorna información del usuario actual basado en el token JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener perfil del usuario autenticado
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...